Geekyshows/python/10.Vector Store Using ChromaDB in LangChain Python/ch9/requirements.txt
langchain==0.3.21
python-decouple==3.8
langchain-openai==0.3.11
langchain-community==0.3.22
chromadb==1.0.3